requirements.txt
asgiref==3.5.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
Django==4.0.1
django-filter==21.1
django-silk==4.2.0
djangorestframework==3.13.1
gprof2dot==2021.2.21
idna==3.3
Jinja2==3.0.3
MarkupSafe==2.0.1
pep517==0.12.0
pip-tools==6.4.0
psycopg2-binary==2.9.3
pycodestyle==2.8.0
Pygments==2.11.2
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.0
urllib3==1.26.8
